requirements.txt
flask
flask_sqlalchemy
qrcode
reportlab
pillow
python-magic
pymysql
openai

